requirements.txt
pandas
openai
streamlit
python-dotenv
